chore: DRY miner
