+++ crates/cast/src/cmd/create2.rs
@@ -3,13 +3,7 @@ use clap::Parser;
 use eyre::{Result, WrapErr};
 use rand::{RngCore, SeedableRng, rngs::StdRng};
 use regex::RegexSetBuilder;
-use std::{
-    sync::{
-        Arc,
-        atomic::{AtomicBool, Ordering},
-    },
-    time::Instant,
-};
+use std::time::Instant;
 
 // https://etherscan.io/address/0x4e59b44847b379578588920ca78fbf26c0b4956c#code
 const DEPLOYER: &str = "0x4e59b44847b379578588920ca78fbf26c0b4956c";
@@ -195,71 +189,27 @@ impl Create2Args {
         sh_println!(
             "Starting to generate deterministic contract address with {n_threads} threads..."
         )?;
-        let mut handles = Vec::with_capacity(n_threads);
-        let found = Arc::new(AtomicBool::new(false));
         let timer = Instant::now();
-
-        // Loops through all possible salts in parallel until a result is found.
-        // Each thread iterates over `(i..).step_by(n_threads)`.
-        for i in 0..n_threads {
-            // Create local copies for the thread.
-            let increment = n_threads;
-            let regex = regex.clone();
-            let regex_len = regex.patterns().len();
-            let found = Arc::clone(&found);
-            handles.push(std::thread::spawn(move || {
-                // Read the first bytes of the salt as a usize to be able to increment it.
-                #[repr(C)]
-                struct B256Aligned(B256, [usize; 0]);
-                let mut salt = B256Aligned(salt, []);
-                // SAFETY: B256 is aligned to `usize`.
-                let salt_word = unsafe {
-                    &mut *salt.0.as_mut_ptr().add(32 - usize::BITS as usize / 8).cast::<usize>()
-                };
-                // Important: add the thread index to the salt to avoid duplicate results.
-                *salt_word = salt_word.wrapping_add(i);
-
-                // Use checksum format only when case_sensitive is enabled.
-                // This avoids an extra keccak256 call per iteration when not needed.
-                let mut checksum_buf = [0u8; 42];
-                let mut hex_buf = [0u8; 40];
-                loop {
-                    // Stop if a result was found in another thread.
-                    if found.load(Ordering::Relaxed) {
-                        break None;
-                    }
-
-                    // Calculate the `CREATE2` address.
-                    #[expect(clippy::needless_borrows_for_generic_args)]
-                    let addr = deployer.create2(&salt.0, &init_code_hash);
-
-                    // Check if the regex matches the calculated address.
-                    // When case_sensitive is true, use EIP-55 checksum format (requires keccak256).
-                    // Otherwise, use lowercase hex to avoid the extra hash computation.
-                    let s = if case_sensitive {
-                        let _ = addr.to_checksum_raw(&mut checksum_buf, None);
-                        // SAFETY: stripping 2 ASCII bytes ("0x") off of an already valid UTF-8
-                        // string is safe.
-                        unsafe { std::str::from_utf8_unchecked(checksum_buf.get_unchecked(2..)) }
-                    } else {
-                        // SAFETY: hex::encode_to_slice always produces valid UTF-8 (hex digits).
-                        let _ = hex::encode_to_slice(addr.as_slice(), &mut hex_buf);
-                        unsafe { std::str::from_utf8_unchecked(&hex_buf) }
-                    };
-                    if regex.matches(s).into_iter().count() == regex_len {
-                        // Notify other threads that we found a result.
-                        found.store(true, Ordering::Relaxed);
-                        break Some((addr, salt.0));
-                    }
-
-                    // Increment the salt for the next iteration.
-                    *salt_word = salt_word.wrapping_add(increment);
-                }
-            }));
-        }
-
-        let results = handles.into_iter().filter_map(|h| h.join().unwrap()).collect::<Vec<_>>();
-        let (address, salt) = results.into_iter().next().unwrap();
+        let regex_len = regex.patterns().len();
+        let mut checksum_buf = [0u8; 42];
+        let mut hex_buf = [0u8; 40];
+        let (address, salt) = super::miner::mine_salt(salt, n_threads, move |salt| {
+            #[expect(clippy::needless_borrows_for_generic_args)]
+            let addr = deployer.create2(&salt, &init_code_hash);
+            // Use checksum format only when case_sensitive is enabled — it requires an extra
+            // keccak256 call, so we fall back to plain hex when case sensitivity is off.
+            let s = if case_sensitive {
+                let _ = addr.to_checksum_raw(&mut checksum_buf, None);
+                // SAFETY: stripping 2 ASCII bytes ("0x") off of an already valid UTF-8 string.
+                unsafe { std::str::from_utf8_unchecked(checksum_buf.get_unchecked(2..)) }
+            } else {
+                // SAFETY: hex::encode_to_slice always produces valid UTF-8 (hex digits).
+                let _ = hex::encode_to_slice(addr.as_slice(), &mut hex_buf);
+                unsafe { std::str::from_utf8_unchecked(&hex_buf) }
+            };
+            (regex.matches(s).into_iter().count() == regex_len).then_some((addr, salt))
+        })
+        .ok_or_else(|| eyre::eyre!("create2 salt mining failed: all threads panicked"))?;
         sh_println!("Successfully found contract address in {:?}", timer.elapsed())?;
         sh_println!("Address: {address}")?;
         sh_println!("Salt: {salt} ({})", U256::from_be_bytes(salt.0))?;
